+++ modules/flowable-engine/src/main/java/org/flowable/engine/delegate/event/FlowableActivityEvent.java
@@ -13,7 +13,6 @@
 package org.flowable.engine.delegate.event;
 
 import org.flowable.engine.common.api.delegate.event.FlowableEvent;
-import org.flowable.engine.impl.delegate.event.FlowableEngineEvent;
 
 /**
  * An {@link FlowableEvent} related to an activity within an execution.
