+++ modules/flowable-engine/src/main/java/org/flowable/engine/delegate/event/FlowableCancelledEvent.java
@@ -12,8 +12,6 @@
  */
 package org.flowable.engine.delegate.event;
 
-import org.flowable.engine.impl.delegate.event.FlowableEngineEvent;
-
 /**
  * @author martin.grofcik
  */
